--- src/system/utils/index.ts
@@ -1,2 +1,0 @@
-export * from './actor';
-export * from './generic';
